// types/enums/Status.ts
export enum Status {
  ACTIVE = "ACTIVE",
  PENDING = "PENDING",
  REJECTED = "REJECTED",
  APPROVED = "APPROVED",
  CANCELLED = "CANCELLED",
  ACCEPTED = "ACCEPTED",
  COMPLETED = "COMPLETED",
  FAILED = "FAILED",
  IN_PROGRESS = "IN_PROGRESS",
  PENDING_PAYMENT = "PENDING_PAYMENT",
  REFUNDED = "REFUNDED",
  REFUND_PENDING = "REFUND_PENDING",
  SCHEDULED = "SCHEDULED",
  ATTENDED = "ATTENDED",
  ABSENT = "ABSENT",
  EXCUSED = "EXCUSED",
  WAITING = "WAITING"
}